+++ index.html
@@ -6,7 +6,7 @@
 </head>
 <body>
 	<div id='mainContainer'></div>
-	<div id='subContainer'></div>
+	<div id='itemContainer'></div>
 	<script type="text/javascript" src="js/bundle.js"></script>
 </body>
 </html>
