+++ js/main.jsx
@@ -1,4 +1,6 @@
 require ('../css/main.css');
-require ('./components/list.jsx');
 var React = require ('react');
 var ReactDOM = require ('react-dom');
+var App = require ('./components/mainApp.jsx');
+var listItem = ('./components/listItem.jsx');
+var Todo = require ('./components/listToDo.jsx');
